down cast  ->discount command

+++ Interfaces/IMenuItem/IMenuItem.h
@@ -5,7 +5,7 @@ class IMenuItem
 public:
 	virtual std::string GetName() const = 0;
 	virtual std::string GetDescription() const = 0;
-	virtual int GetPrice() const = 0;
+	virtual double GetPrice() const = 0;
 	virtual std::string GetSize() const = 0;
 	virtual int GetQuantity() const = 0;
 	virtual void Display() const = 0;
